move multibody plotting function to plots directory
